--- a/torch/optim/adagrad.py
+++ b/torch/optim/adagrad.py
@@ -321,7 +321,11 @@ def _multi_tensor_adagrad(
     grouped_tensorlists = Optimizer._group_tensors_by_device_and_dtype([params, grads, state_sums, state_steps])
     for ((device_params, device_grads, device_state_sums, device_state_steps), _) in grouped_tensorlists.values():
 
-        device_has_sparse_grad = any(grad.is_sparse for grad in device_grads)
+        device_has_sparse_grad = False
+        for grad in device_grads:
+            if grad.is_sparse:
+                device_has_sparse_grad = True
+                break
 
         if device_has_sparse_grad:
             return _single_tensor_adagrad(