--- a/torch/optim/adam.py
+++ b/torch/optim/adam.py
@@ -488,12 +488,12 @@ def _multi_tensor_adam(params: List[Tensor],
         # Handle complex parameters
         for i in range(len(device_params)):
             if torch.is_complex(device_params[i]):
+                device_params[i] = torch.view_as_real(device_params[i])
                 device_grads[i] = torch.view_as_real(device_grads[i])
                 device_exp_avgs[i] = torch.view_as_real(device_exp_avgs[i])
                 device_exp_avg_sqs[i] = torch.view_as_real(device_exp_avg_sqs[i])
                 if len(device_max_exp_avg_sqs) > 0:
                     device_max_exp_avg_sqs[i] = torch.view_as_real(device_max_exp_avg_sqs[i])
-                device_params[i] = torch.view_as_real(device_params[i])
 
         # update steps
         torch._foreach_add_(device_state_steps, 1)