@jon-chuang 2023-10-05
Actually, I know how to speed this up even further. We can add a `has_complex` case, similar to `has_sparse_grad`