@jon-chuang 2023-10-05
@janeyx99 to be more explicit, this is how single_tensor case always shortcuts:
https://github.com/pytorch/pytorch/blob/0296632b9266f1958d7194b99116e392be4a3705/torch/optim/adam.py#L375
https://github.com/pytorch/pytorch/blob/0296632b9266f1958d7194b99116e392be4a3705/torch/optim/adamax.py#L262
https://github.com/pytorch/pytorch/blob/0296632b9266f1958d7194b99116e392be4a3705/torch/optim/adamw.py#L397
https://github.com/pytorch/pytorch/blob/0296632b9266f1958d7194b99116e392be4a3705/torch/optim/asgd.py#L251
https://github.com/pytorch/pytorch/blob/0296632b9266f1958d7194b99116e392be4a3705/torch/optim/rmsprop.py#L275
https://github.com/pytorch/pytorch/blob/0296632b9266f1958d7194b99116e392be4a3705/torch/optim/rprop.py#L238